+++ ethereum/permissioning/build.gradle
@@ -45,6 +45,7 @@ dependencies {
   implementation 'io.consensys.tuweni:tuweni-bytes'
   implementation 'io.consensys.tuweni:tuweni-toml'
   implementation 'io.consensys.tuweni:tuweni-units'
+  implementation 'javax.inject:javax.inject'
   implementation 'org.web3j:abi'
 
   testImplementation project(':config')
