+++ ethereum/permissioning/src/main/java/org/hyperledger/besu/ethereum/permissioning/pluginadapter/PermissioningServiceImpl.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.services;
+package org.hyperledger.besu.ethereum.permissioning.pluginadapter;
 
 import org.hyperledger.besu.plugin.services.PermissioningService;
 import org.hyperledger.besu.plugin.services.permissioning.NodeConnectionPermissioningProvider;
