+++ platform/build.gradle
@@ -59,6 +59,7 @@ dependencies {
     api project(':ethereum:eth')
     api project(':ethereum:p2p')
     api project(':ethereum:permissioning')
+    api project(':ethereum:permissioning:plugin-api')
     api project(':ethereum:referencetests')
     api project(':ethereum:rlp')
     api project(':ethereum:trie')
