+++ settings.gradle
@@ -46,6 +46,7 @@ include 'ethereum:evmtool'
 include 'ethereum:mock-p2p'
 include 'ethereum:p2p'
 include 'ethereum:permissioning'
+include 'ethereum:permissioning:plugin-api'
 include 'ethereum:referencetests'
 include 'ethereum:rlp'
 include 'ethereum:ethstats'
